+++ CHANGELOG.md
@@ -76,6 +76,8 @@ For instructions on upgrading to newer versions, visit
 
 * \#1668 Ensure Mongoid logger exists before calling warn. (Rémy Coutable)
 
+* \#1661 Ensure uniqueness validation works on cloned documents.
+
 * \#1659 Clear delayed atomic sets when resetting the same embedded relation.
 
 * \#1656/#1657 Don't hit database for uniqueness validation if BOTH scope